backend/requirements.txt
flask
pymongo
flask-bcrypt
python-dotenv